🎨 Model 파일 분리
